parser: Fix EXECVE for very long commmand lines
argv can be longer than 2^16 elements. Because the number part
couldn't be parsed, those arguments would presumably be turned into
regular keys.

Changing the value from u16 to u32 fixes this. The fallback function
for regular key/value pairs is removed.

+++ src/types.rs
@@ -261,9 +261,9 @@ pub enum Key {
     /// translated / "enriched" values
     NameTranslated(NVec),
     /// `a0`, `a1`, `a2[0]`, `a2[1]`…
-    Arg(u16, Option<u16>),
+    Arg(u32, Option<u16>),
     /// `a0_len` …
-    ArgLen(u16),
+    ArgLen(u32),
     Literal(&'static str),
 }
